refactor : step3 리뷰 사항 반영2
* domain 서비스 domain 패키지로 이동
