+++ src/main/java/kitchenpos/table/controller/TableGroupRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.table.controller;
 
 import java.net.URI;
-import kitchenpos.table.application.TableGroupBo;
+import kitchenpos.table.domain.TableGroupBo;
 import kitchenpos.table.domain.TableGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
