+++ src/test/java/kitchenpos/bo/MenuGroupBoTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.bo;
 
 import kitchenpos.menu.domain.MenuGroupDao;
-import kitchenpos.menu.application.MenuGroupBo;
+import kitchenpos.menu.domain.MenuGroupBo;
 import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
