+++ src/test/java/kitchenpos/bo/OrderBoTest.java
@@ -3,10 +3,10 @@ package kitchenpos.bo;
 import kitchenpos.menu.domain.MenuDao;
 import kitchenpos.order.domain.OrderDao;
 import kitchenpos.order.domain.OrderLineItemDao;
-import kitchenpos.table.application.OrderTableDao;
+import kitchenpos.table.domain.OrderTableDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.application.OrderBo;
+import kitchenpos.order.domain.OrderBo;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
