+++ src/test/java/kitchenpos/controller/TableGroupRestControllerTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.controller;
 
-import kitchenpos.table.application.TableGroupBo;
+import kitchenpos.table.domain.TableGroupBo;
 import kitchenpos.table.domain.TableGroup;
 import kitchenpos.table.controller.TableGroupRestController;
 import org.junit.jupiter.api.Test;
